+++ app/src/main/java/com/dvinc/notepad/data/repository/note/NoteDataRepository.kt
@@ -32,7 +32,7 @@ class NoteDataRepository
                 }
     }
 
-    override fun deleteNoteById(id: Int): Completable {
+    override fun deleteNoteById(id: Long): Completable {
         return Completable.fromAction { noteDao.deleteNoteById(id) }
     }
